requirements.txt
Flask
Flask-MySQL
Flask-SQLAlchemy
Flask-WTF
pymysql
gunicorn
requests
RPi.GPIO
flask-cors
